cmd/rescached: make loading configuration file as optional
Previously, if configuration file is not defined it will be default
to "/etc/rescached/rescached.cfg" and if the file is not exist it will
return an error with nil options.

This changes, make the configuration file to be an optional.  If its not
exist or error when parsing options, rescached will set it with default
values.