cmd/rescached: remove documentation on rescached.cfg
Its hard to keep the same documentation in two different places to be
always in sync.

The documentation for configuration is located in two
places: in the configuration file itself and in its manual
rescached.cfg(5).

This changes remove the documentation in configuration and update the
documentation in man file only.