all: refactor the APIs to manage hosts.d resource records (RR)
There are two APIs to manage hosts.d RR: create and delete.

Previously, the create API pass the hosts name inside and values inside
the path,

  POST /hosts.d/<name>/rr?domain=&value=

This commit changes the request type to form, so all parameters move
to body,

  POST /hosts.d/rr
  Content-Type: application/x-www-form-urlencoded

  name=&domain=&value=

On delete API, we changes the name parameter to be send on query
parameter along with domain to be deleted.
Previously, the delete API was

  DELETE /hosts.d/<name>/rr?domain=

now it become

  DELETE /hosts.d/rr?name=&domain=