all: refactoring HTTP API for adding new zone.d record
Previously, the request for adding new record on zone file is by passing
the zone name and type inside the path,

  /zone.d/:name/rr/:type

This commit changes it to pass all parameters inside the request body
as JSON,

	{
		"zone": <string>,
		"kind": <string>,
		"record": <base64 string|base64 JSON>
	}

For example, to add A record for subdomain "www" to zone file "my.zone",
the request format would be,

	{
		"zone": "my.zone",
		"kind": "A",
		"record": "eyJOYW1lIjoid3d3IiwiVmFsdWUiOiIxMjcuMC4wLjEifQ=="
	}

Where "record" value is equal to `{"Name":"www","Value":"127.0.0.1"}`.

On the response, we changes it to return only the new record instead of
all record in the zone.