cmd/resolver: implement command to add new record to zone
The command has the following signature,

  zone.d rr add <zone> <"@" | subdomain> <ttl> <type> <class> <value> ...::

The domain name can be set to origin using "@" or empty string, subdomain
(without ending with "."), or fully qualified domain name (end with ".").

If ttl is set to 0, it will default to 604800 (7 days).

List of valid type are A, NS, CNAME, PTR, MX, TXT, and AAAA.

List of valid class are IN, CS, HS.