all: export and rename the hostsBlock type to Blockd
In the Environment, we have a field HostsBlocks that reference
an unexported type hostsFile.
In order for any package to be able to use this field, we need to
export this type.

While at it, we rename the type from hostsBlock to Blockd to make it
simple and consistent with name.